Start using CounterContext exported by slate-counters
