+++ src/screens/app/Draft/plugins/Xref/index.tsx
@@ -1,195 +1,15 @@
-import * as React from 'react'
-import * as PropTypes from 'prop-types'
 import { ReactLocalization } from 'fluent-react/compat'
-import { connect } from 'react-redux'
 import { Map } from 'immutable'
-import { Block, Inline, Value } from 'slate'
-import { Plugin, RenderInlineProps } from 'slate-react'
+import { Inline, Value } from 'slate'
+import { Plugin } from 'slate-react'
 
-import { State } from 'src/store/reducers'
-import { ReferenceTarget } from 'src/store/types'
+import renderInline from './render'
 
 const XrefPlugin: Plugin = {
-  renderInline(props, _, next) {
-    if (props.node.type === 'xref') {
-      return <Xref {...props} />
-    }
-
-    return next()
-  },
+  renderInline,
 }
 
 export default XrefPlugin
-
-interface XrefProps extends RenderInlineProps {
-  referenceTargets?: ReferenceTarget[]
-}
-
-const mapStateTopProps = (
-  { modules: { referenceTargets } }: State,
-  { node }: RenderInlineProps
-) => {
-  const document = node.data.get('document')
-  if (document) {
-    return {
-      referenceTargets: referenceTargets.get(document),
-    }
-  }
-  return {}
-}
-
-type RefTranslation = [ReactLocalization, string, object, string]
-
-const Xref = connect(mapStateTopProps)(class Xref extends React.Component<XrefProps> {
-  static contextTypes = {
-    counters: PropTypes.instanceOf(Map as any),
-    l10n: PropTypes.instanceOf(ReactLocalization),
-    uiL10n: PropTypes.instanceOf(ReactLocalization),
-  }
-
-  componentDidMount() {
-    const { l10n, uiL10n } = this.context
-
-    if (l10n) l10n.subscribe(this)
-    if (uiL10n) uiL10n.subscribe(this)
-  }
-
-  componentWillUnmount() {
-    const { l10n, uiL10n } = this.context
-
-    if (l10n) l10n.unsubscribe(this)
-    if (uiL10n) uiL10n.unsubscribe(this)
-  }
-
-  relocalize() {
-    this.forceUpdate()
-  }
-
-  render() {
-    const { node, attributes } = this.props
-    const { uiL10n } = this.context
-
-    const targetDocument = node.data.get('document')
-
-    const [l10n, l10nKey, args, href] = targetDocument
-      ? this.renderRemote()
-      : this.renderLocal()
-
-    const title = uiL10n.getString('editor-tools-xref-hover-tooltip')
-    const text = l10n.getString(l10nKey, args)
-
-    return (
-      <a
-        href={href}
-        onClick={this.onClick}
-        title={title}
-        {...attributes}
-      >
-        {text}
-      </a>
-    )
-  }
-
-  renderLocal(): RefTranslation {
-    const { node, editor } = this.props
-    const { counters, l10n, uiL10n } = this.context
-    const targetKey = node.data.get('target')
-    const target = editor.value.document.getNode(targetKey) as unknown as Block | Inline
-
-    let l10nKey
-    const args: {
-      case: string
-      note?: string
-      [key: string]: any
-    } = { case: node.data.get('case') }
-    let localization
-
-    if (target) {
-      switch (target.type) {
-      case 'admonition':
-        l10nKey = 'xref-label-' + target.data.get('type')
-        break
-
-      default:
-        l10nKey = 'xref-label-' + target.type
-        break
-      }
-
-      localization = l10n
-
-      const cnts = counters.get(targetKey) || Map()
-      for (const [name, value] of cnts) {
-        // Temporary solution until we will support all types of notes
-        if (name === 'admonition') {
-          args.note = value
-        } else {
-          args[name] = value
-        }
-      }
-    } else {
-      console.warn(`Undefined target in ${node.key}: ${targetKey}`)
-      l10nKey = 'editor-tools-xref-label-local-reference-missing'
-      localization = uiL10n
-    }
-
-    return [localization, l10nKey, args, "#" + targetKey]
-  }
-
-  renderRemote(): RefTranslation {
-    const { node, referenceTargets } = this.props
-    const { l10n, uiL10n } = this.context
-
-    const targetDocument = node.data.get('document')
-    const targetKey = node.data.get('target')
-
-    const findTarget = (refs: ReferenceTarget[] = []): ReferenceTarget | undefined => {
-      let result
-      refs.some(target => {
-        if (target.id === targetKey) {
-          result = target
-          return true
-        }
-        if (target.children.length) {
-          const res = findTarget(target.children)
-          if (res) {
-            result = res
-            return true
-          }
-        }
-        return false
-      })
-      return result
-    }
-    const target = findTarget(referenceTargets)
-
-    let l10nKey
-    const args = { case: node.data.get('case') }
-    let localization
-
-    if (!referenceTargets) {
-      l10nKey = 'editor-tools-xref-label-remote-loading'
-      localization = uiL10n
-    } else if (target) {
-      l10nKey = 'xref-label-' + target.type
-      args[target.type] = target.counter
-      localization = l10n
-    } else {
-      console.warn(`Undefined target in ${node.key}: ${targetKey} from document ${targetDocument}`)
-      l10nKey = 'editor-tools-xref-label-remote-reference-missing'
-      localization = uiL10n
-    }
-
-    const href = `/modules/${targetDocument}#${targetKey}`
-
-    return [localization, l10nKey, args, href]
-  }
-
-  onClick = (ev: React.MouseEvent<HTMLAnchorElement>) => {
-    if (!ev.ctrlKey) {
-      ev.preventDefault()
-    }
-  }
-})
 
 export function collectForeignDocuments(value: Value): string[] {
   const documents = new Set<string>()
